Move RuleExecutionMetrics to core
